Code fixes 3

+++ qrgui/mainwindow/mainWindow.h
@@ -324,7 +324,7 @@ private:
 	qReal::gui::PaletteTree *mPaletteTree;
 	FindManager *mFindHelper;
 
-	size_t mDiagramNumber;
+	int mDiagramNumber;
 
 };
